Refactor OptimizeIndexing to make it more extensible
This PR decouples the logic necessary to compute bounds on variables
from the logic that uses this info to perform the strenght analysis on
int64 variables. While doing so, it tries to minimize the number of
attributes of the class in favour of local variables.

This class is now accessible from any `LoopBody` object.

ghstack-source-id: 63d2c7735ddaa6455222fbc1fd24af00e99ffc0e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/100549