[inductor] Add single pass "var_unnormalized" reduction_type
This is a bit inefficient because it computes the mean and throws it
away since ir.Reduction nodes only have 1 output. However, the mean
can at least be scheduled into the same loop as the variance now since
there is no data dependency. Thus we can take fewer passes over the
data.

ghstack-source-id: 095ed04d182b8650ac2dd77e490a1c27915de39d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/102486